Updated package.json to watch for *.spec Reverted testing_util.js refactoring
