+++ package.json
@@ -8,7 +8,7 @@
         "url": "https://github.com/reflexdemon/slush-angular-gulp"
     },
     "scripts": {
-        "test": "NODE_ENV=test mocha -R spec test/*_test.js"
+        "test": "NODE_ENV=test mocha -R spec test/*.spec.js"
     },
     "keywords": [
         "slushgenerator",
